Apply solid principles
